+++ test/kura_test_repo.erl
@@ -28,6 +28,7 @@ otp_app() -> kura.
 
 start() ->
     application:set_env(pg_types, uuid_format, string),
+    application:set_env(kura, dialect, kura_dialect_pg),
     application:set_env(kura, kura_test_repo, #{
         pool => kura_test_repo,
         database => <<"kura_test">>,